+++ src/functions.php
@@ -1,0 +1,423 @@
+<?php
+
+namespace Pramda;
+
+function add()
+{
+    return call_user_func_array(['P', 'add'], func_get_args());
+}
+
+function all()
+{
+    return call_user_func_array(['P', 'all'], func_get_args());
+}
+
+function allPass()
+{
+    return call_user_func_array(['P', 'allPass'], func_get_args());
+}
+
+function any()
+{
+    return call_user_func_array(['P', 'any'], func_get_args());
+}
+
+function append()
+{
+    return call_user_func_array(['P', 'append'], func_get_args());
+}
+
+function appendTo()
+{
+    return call_user_func_array(['P', 'appendTo'], func_get_args());
+}
+
+function apply()
+{
+    return call_user_func_array(['P', 'apply'], func_get_args());
+}
+
+function chain()
+{
+    return call_user_func_array(['P', 'chain'], func_get_args());
+}
+
+function compose()
+{
+    return call_user_func_array(['P', 'compose'], func_get_args());
+}
+
+function concat()
+{
+    return call_user_func_array(['P', 'concat'], func_get_args());
+}
+
+function contains()
+{
+    return call_user_func_array(['P', 'contains'], func_get_args());
+}
+
+function converge()
+{
+    return call_user_func_array(['P', 'converge'], func_get_args());
+}
+
+function countBy()
+{
+    return call_user_func_array(['P', 'countBy'], func_get_args());
+}
+
+function curry2()
+{
+    return call_user_func_array(['P', 'curry2'], func_get_args());
+}
+
+function curry3()
+{
+    return call_user_func_array(['P', 'curry3'], func_get_args());
+}
+
+function curry4()
+{
+    return call_user_func_array(['P', 'curry4'], func_get_args());
+}
+
+function curryN()
+{
+    return call_user_func_array(['P', 'curryN'], func_get_args());
+}
+
+function dec()
+{
+    return call_user_func_array(['P', 'dec'], func_get_args());
+}
+
+function divide()
+{
+    return call_user_func_array(['P', 'divide'], func_get_args());
+}
+
+function each()
+{
+    return call_user_func_array(['P', 'each'], func_get_args());
+}
+
+function eq()
+{
+    return call_user_func_array(['P', 'eq'], func_get_args());
+}
+
+function eqBy()
+{
+    return call_user_func_array(['P', 'eqBy'], func_get_args());
+}
+
+function file()
+{
+    return call_user_func_array(['P', 'file'], func_get_args());
+}
+
+function filter()
+{
+    return call_user_func_array(['P', 'filter'], func_get_args());
+}
+
+function flatten()
+{
+    return call_user_func_array(['P', 'flatten'], func_get_args());
+}
+
+function flip()
+{
+    return call_user_func_array(['P', 'flip'], func_get_args());
+}
+
+function gt()
+{
+    return call_user_func_array(['P', 'gt'], func_get_args());
+}
+
+function gte()
+{
+    return call_user_func_array(['P', 'gte'], func_get_args());
+}
+
+function head()
+{
+    return call_user_func_array(['P', 'head'], func_get_args());
+}
+
+function identity()
+{
+    return call_user_func_array(['P', 'identity'], func_get_args());
+}
+
+function inc()
+{
+    return call_user_func_array(['P', 'inc'], func_get_args());
+}
+
+function join()
+{
+    return call_user_func_array(['P', 'join'], func_get_args());
+}
+
+function last()
+{
+    return call_user_func_array(['P', 'last'], func_get_args());
+}
+
+function lt()
+{
+    return call_user_func_array(['P', 'lt'], func_get_args());
+}
+
+function lte()
+{
+    return call_user_func_array(['P', 'lte'], func_get_args());
+}
+
+function map()
+{
+    return call_user_func_array(['P', 'map'], func_get_args());
+}
+
+function mathMod()
+{
+    return call_user_func_array(['P', 'mathMod'], func_get_args());
+}
+
+function max()
+{
+    return call_user_func_array(['P', 'max'], func_get_args());
+}
+
+function maxBy()
+{
+    return call_user_func_array(['P', 'maxBy'], func_get_args());
+}
+
+function merge()
+{
+    return call_user_func_array(['P', 'merge'], func_get_args());
+}
+
+function mergeAll()
+{
+    return call_user_func_array(['P', 'mergeAll'], func_get_args());
+}
+
+function min()
+{
+    return call_user_func_array(['P', 'min'], func_get_args());
+}
+
+function minBy()
+{
+    return call_user_func_array(['P', 'minBy'], func_get_args());
+}
+
+function modulo()
+{
+    return call_user_func_array(['P', 'modulo'], func_get_args());
+}
+
+function multiply()
+{
+    return call_user_func_array(['P', 'multiply'], func_get_args());
+}
+
+function nd()
+{
+    return call_user_func_array(['P', 'nd'], func_get_args());
+}
+
+function negate()
+{
+    return call_user_func_array(['P', 'negate'], func_get_args());
+}
+
+function nth()
+{
+    return call_user_func_array(['P', 'nth'], func_get_args());
+}
+
+function of()
+{
+    return call_user_func_array(['P', 'of'], func_get_args());
+}
+
+function partial()
+{
+    return call_user_func_array(['P', 'partial'], func_get_args());
+}
+
+function partialRight()
+{
+    return call_user_func_array(['P', 'partialRight'], func_get_args());
+}
+
+function partition()
+{
+    return call_user_func_array(['P', 'partition'], func_get_args());
+}
+
+function pipe()
+{
+    return call_user_func_array(['P', 'pipe'], func_get_args());
+}
+
+function pluck()
+{
+    return call_user_func_array(['P', 'pluck'], func_get_args());
+}
+
+function prepend()
+{
+    return call_user_func_array(['P', 'prepend'], func_get_args());
+}
+
+function prependTo()
+{
+    return call_user_func_array(['P', 'prependTo'], func_get_args());
+}
+
+function product()
+{
+    return call_user_func_array(['P', 'product'], func_get_args());
+}
+
+function prop()
+{
+    return call_user_func_array(['P', 'prop'], func_get_args());
+}
+
+function propOf()
+{
+    return call_user_func_array(['P', 'propOf'], func_get_args());
+}
+
+function propOr()
+{
+    return call_user_func_array(['P', 'propOr'], func_get_args());
+}
+
+function props()
+{
+    return call_user_func_array(['P', 'props'], func_get_args());
+}
+
+function reduce()
+{
+    return call_user_func_array(['P', 'reduce'], func_get_args());
+}
+
+function reverse()
+{
+    return call_user_func_array(['P', 'reverse'], func_get_args());
+}
+
+function set()
+{
+    return call_user_func_array(['P', 'set'], func_get_args());
+}
+
+function size()
+{
+    return call_user_func_array(['P', 'size'], func_get_args());
+}
+
+function slice()
+{
+    return call_user_func_array(['P', 'slice'], func_get_args());
+}
+
+function sort()
+{
+    return call_user_func_array(['P', 'sort'], func_get_args());
+}
+
+function split()
+{
+    return call_user_func_array(['P', 'split'], func_get_args());
+}
+
+function subtract()
+{
+    return call_user_func_array(['P', 'subtract'], func_get_args());
+}
+
+function sum()
+{
+    return call_user_func_array(['P', 'sum'], func_get_args());
+}
+
+function tail()
+{
+    return call_user_func_array(['P', 'tail'], func_get_args());
+}
+
+function take()
+{
+    return call_user_func_array(['P', 'take'], func_get_args());
+}
+
+function takeLast()
+{
+    return call_user_func_array(['P', 'takeLast'], func_get_args());
+}
+
+function takeWhile()
+{
+    return call_user_func_array(['P', 'takeWhile'], func_get_args());
+}
+
+function toArray()
+{
+    return call_user_func_array(['P', 'toArray'], func_get_args());
+}
+
+function trampoline()
+{
+    return call_user_func_array(['P', 'trampoline'], func_get_args());
+}
+
+function unapply()
+{
+    return call_user_func_array(['P', 'unapply'], func_get_args());
+}
+
+function unary()
+{
+    return call_user_func_array(['P', 'unary'], func_get_args());
+}
+
+function uniq()
+{
+    return call_user_func_array(['P', 'uniq'], func_get_args());
+}
+
+function unnest()
+{
+    return call_user_func_array(['P', 'unnest'], func_get_args());
+}
+
+function values()
+{
+    return call_user_func_array(['P', 'values'], func_get_args());
+}
+
+function zip()
+{
+    return call_user_func_array(['P', 'zip'], func_get_args());
+}
+
+function zipAssoc()
+{
+    return call_user_func_array(['P', 'zipAssoc'], func_get_args());
+}
+
+function zipWith()
+{
+    return call_user_func_array(['P', 'zipWith'], func_get_args());
+}